Fix remaining -Wextra warnings (r15 round)
- cluster_validation_utils.cpp: restore from regex corruption, add
  .named_compile_args to all EthernetConfig initializers
- tt_metal.cpp: cast DataMovementProcessor underlying type to uint32_t
- core_assignment.cpp: use uint32_t for y_coord, x_coord, num_dram_banks loops
- shape.cpp: cast rank to uint64_t in normalized_index comparison

+++ tt_metal/common/shape.cpp
@@ -50,7 +50,7 @@ uint32_t Shape::get_normalized_index(std::int64_t index) const {
     std::int64_t rank = static_cast<std::int64_t>(this->rank());
     std::uint64_t normalized_index = index >= 0 ? index : rank + index;
     TT_FATAL(
-        normalized_index < rank,
+        normalized_index < static_cast<uint64_t>(rank),
         "Index is out of bounds for the rank, should be between 0 and {} however is {}",
         rank - 1,
         normalized_index);
